ai_streamer/requirements.txt
# AI虚拟主播项目依赖
# 安装: pip install -r requirements.txt

# 核心依赖
anthropic>=0.21.0
playwright>=1.42.0
websockets>=12.0
aiohttp>=3.9.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# TTS语音合成
edge-tts>=6.1.0
# GPT-SoVITS需要单独安装，参考: https://github.com/RVC-Boss/GPT-SoVITS

# VTube Studio控制
python-osc>=1.8.3

# 工具库
requests>=2.31.0
pillow>=10.2.0
numpy>=1.26.0
loguru>=0.7.0

# 开发依赖(可选)
pytest>=8.0.0
black>=24.0.0
